bugfix: deadlock in L67:cache.go:FetchTweets()
The parent goroutine adds an empty struct{} to the 'fetchers'
channel, which has a buffer size of maxfetchers. When the
fetchers channel is full, the parent goroutine will block
on send until the channel empties. However, as the fetchers
channel is read from in the same goroutine, rather than the
child goroutines, the same thread of execution is expected
to read from the channel as is blocking on send. This will
block until the program is killed. As maxfetchers is set
to a sufficiently high value, this will not manifest unless
one has many people to follow and, as a result, over 50
twtxt.txt files to fetch at once.

Moving the fetchers channel read for each loop iteration
to the child goroutine allows the channel to be read on
a separate thread of execution from the one blocking on
send, preventing this deadlock while achieving the desired
ratelimiting behavior.