Revert "bugfix: deadlock in L67:cache.go:FetchTweets()"
This reverts commit 8e762fc8a598c4bfd86d79e6bcb59e6eea10d28f.

This resulted in the HTTP-fetching goroutine hanging on
`tweetsch <- tweets` -- since the parent routine in turn hangs on
`fetchers <- struct{}{}` instead of spinning up all goroutines at once
and beginning to read from the `tweetsch' chan. Given how the
fetching-goroutine currently relies on the main routine reading from
that channel, this is not workable.